app/requirement.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==1.10.9
python-multipart==0.0.6
pillow==10.0.0
pytesseract==0.3.10
pdf2image==1.16.3
